Rails 7.0 Compatibility (#1365)
* Make paper_trail work with Rails 7.0

* from class_methods do back to module ClassMethods

* add spec for PostgresArraySerializer to boost coverage

* lint the spec for PostgresArraySerializer

* lint the spec for PostgresArraySerializer again

* and now make that linted spec pass again

* test object change scopes a bit

* round out json and jsonb testing of object scopes

* test some other code paths to increase coverage

* linting

* linting

* mess with yaml loading in test

* oddball cop for double quotes

* use Rails public API for compatibility rather than instance_variable_set

+++ spec/dummy_app/db/migrate/20110208155312_set_up_test_tables.rb
@@ -131,10 +131,10 @@ class SetUpTestTables < ::ActiveRecord::Migration::Current
       %w[json jsonb].each do |j|
         table_name = j + "_versions"
         create_table table_name, force: true do |t|
-          t.string :item_type, null: false
-          t.integer :item_id, null: false
-          t.string :event, null: false
-          t.string :whodunnit
+          t.string   :item_type, null: false
+          t.bigint   :item_id, null: false
+          t.string   :event, null: false
+          t.string   :whodunnit
           t.public_send j, :object
           t.public_send j, :object_changes
           t.datetime :created_at, limit: 6
